Anglicise aide et restructure les fichiers de metadonnées
